Fix the findings of the adversarial multi-agent review
Measurement validity:
- disp for the om vptr rows no longer subtracts the nvf baseline: their timed
  region never dereferences the receiver, so the subtraction credited a
  ~270-cycle cold miss they never paid, understating cold dispatch ~2x. Rows
  declare touches_receiver; for those that do not, disp = net.
- clflush now flushes the C++ v-table from its head (vptr-16), not the address
  point: the type_info slot at vptr-8 escaped eviction whenever the vptr was
  64-byte aligned, a compiler-asymmetric hole.
- verify() checks every path against oracles computed from the class contracts
  alone (x + tag, the tag rule, x + b.tag for the dd yardstick) instead of
  against another path under test, which passed when every registry was wrong
  the same way.

Tooling:
- matrix.sh clears stale run directories (committed results made mixing the
  norm) and no longer pipes the clflush leg, whose failures tail swallowed.
- report.py: clear error on truncated CSVs, no crash below three passes, and
  captions count the CSVs actually loaded; every section states its pass count.
- CLI: missing values and zero/overflowing --objects/--reps/--sweep-mb are
  rejected instead of segfaulting or silently measuring nothing.
- calibrate_tsc_hz: int64 arithmetic; 32-bit long overflowed on a >2.15 s
  stall, empirically yielding a 56 GHz "calibration".

Results regenerated (7 passes, committed registration) and every inline figure
re-derived; stale figures and repository claims in both docs corrected.

+++ src/timing.hpp
@@ -200,11 +200,15 @@ inline auto calibrate_tsc_hz(double seconds = 0.2) -> double {
     clock_gettime(CLOCK_MONOTONIC, &t0);
     auto c0 = tsc_start();
 
-    const auto target = static_cast<long>(seconds * 1e9);
+    // int64 throughout: on -m32, `long` is 32 bits and a stall of ~2.15 s
+    // during the spin (SIGSTOP, hypervisor descheduling) made the multiply
+    // overflow -- UB, and an absurd calibrated frequency.
+    const auto target = static_cast<std::int64_t>(seconds * 1e9);
 
     for (;;) {
         clock_gettime(CLOCK_MONOTONIC, &t1);
-        auto so_far = (t1.tv_sec - t0.tv_sec) * 1000000000L +
+        auto so_far = static_cast<std::int64_t>(t1.tv_sec - t0.tv_sec) *
+                1000000000 +
             (t1.tv_nsec - t0.tv_nsec);
 
         if (so_far >= target) {
@@ -214,7 +218,8 @@ inline auto calibrate_tsc_hz(double seconds = 0.2) -> double {
 
     auto c1 = tsc_stop();
     auto ns = static_cast<double>(
-        (t1.tv_sec - t0.tv_sec) * 1000000000L + (t1.tv_nsec - t0.tv_nsec));
+        static_cast<std::int64_t>(t1.tv_sec - t0.tv_sec) * 1000000000 +
+        (t1.tv_nsec - t0.tv_nsec));
 
     return static_cast<double>(elapsed(c0, c1)) * 1e9 / ns;
 }
